+++ src/lib/vega-wallet/vega-wallet-service.ts
@@ -1,6 +1,6 @@
 import { VoteValue } from "../../__generated__/globalTypes";
 import { VegaKey } from "../../contexts/app-state/app-state-context";
-import { VOTE_VALUE_MAP } from "../../routes/governance/vote-types";
+import { VOTE_VALUE_MAP } from "../../routes/governance/components/vote-details";
 import { LocalStorage } from "../storage";
 import { GenericErrorResponse } from "./vega-wallet-types";
