+++ src/routes/governance/components/current-proposal-status/current-proposal-status.tsx
@@ -3,9 +3,9 @@ import "./current-proposal-status.scss";
 import { formatDistanceToNow } from "date-fns";
 import { useTranslation } from "react-i18next";
 
-import { ProposalState } from "../../__generated__/globalTypes";
-import { Proposals_proposals } from "./__generated__/Proposals";
-import { useVoteInformation } from "./hooks";
+import { ProposalState } from "../../../../__generated__/globalTypes";
+import { Proposals_proposals } from "../../__generated__/Proposals";
+import { useVoteInformation } from "../../hooks";
 
 export const CurrentProposalStatus = ({
   proposal,
