+++ src/routes/governance/components/current-proposal-status/index.tsx
@@ -1,0 +1,1 @@
+export { CurrentProposalStatus } from "./current-proposal-status";
