+++ src/routes/governance/components/proposal-change-table/index.tsx
@@ -1,0 +1,1 @@
+export { ProposalChangeTable } from "./proposal-change-table";
