+++ src/routes/governance/components/vote-details/index.tsx
@@ -1,0 +1,2 @@
+export { VoteDetails } from "./vote-details";
+export { VOTE_VALUE_MAP } from "./vote-types";
