+++ src/routes/governance/components/vote-details/vote-buttons.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors";
+@import "../../../../styles/colors";
 
 .vote-buttons {
   &__button-container {
