+++ src/routes/governance/components/vote-details/vote-details.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors";
+@import "../../../../styles/colors";
 
 .vote-details {
   &__vote-buttons {
