+++ src/routes/governance/components/vote-details/vote-progress.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors';
+@import "../../../../styles/colors";
 
 .vote-progress {
   &__indicator {
