+++ src/routes/governance/components/vote-details/vote-progress.tsx
@@ -1,6 +1,6 @@
 import "./vote-progress.scss";
 
-import { BigNumber } from "../../lib/bignumber";
+import { BigNumber } from "../../../../lib/bignumber";
 
 export const VoteProgress = ({
   progress,
