+++ src/routes/governance/components/vote-details/vote-types.ts
@@ -1,4 +1,4 @@
-import { VoteValue } from "../../__generated__/globalTypes";
+import { VoteValue } from "../../../../__generated__/globalTypes";
 
 export const VOTE_VALUE_MAP = {
   [VoteValue.Yes]: "VALUE_YES",
