--- src/routes/governance/current-proposal-status.scss
@@ -1,11 +1,0 @@
-@import '../../styles/colors';
-
-.current-proposal-status {
-  &__pass {
-    color: $vega-green3
-  }
-
-  &__fail {
-    color: $vega-red3
-  }
-}
