+++ src/routes/governance/hooks/index.tsx
@@ -1,0 +1,1 @@
+export { useVoteInformation } from "./use-vote-information";
