--- src/routes/governance/proposal-change-text.scss
@@ -1,11 +1,0 @@
-@import "../../styles/colors";
-
-.proposal-change-text {
-  code {
-    display: inline-block;
-    padding: 0px 4px;
-    border-radius: 3px;
-    color: $white;
-    background-color: $dark-gray5;
-  }
-}
