--- src/routes/governance/proposal-change-text.tsx
@@ -1,29 +1,0 @@
-import "./proposal-change-text.scss";
-
-import { Trans } from "react-i18next";
-
-interface ProposalChangeTextProps {
-  networkParam: {
-    key: string;
-    value: string;
-  };
-}
-
-export const ProposalChangeText = ({
-  networkParam,
-}: ProposalChangeTextProps) => {
-  return (
-    <span className="proposal-change-text">
-      <Trans
-        i18nKey="proposalChange"
-        values={{
-          key: networkParam.key,
-          value: networkParam.value,
-        }}
-        components={{
-          code: <code />,
-        }}
-      />
-    </span>
-  );
-};
